add: add UI for student management
